refactor: そのページのみで使用しているコンポーネントをpagesに移行
